requirements.txt
openai==1.40.0
termcolor
fastapi
uvicorn
python-multipart
jinja2
python-dotenv